pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gitpython>=3.1.44",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "openai>=1.77.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "slack-sdk>=3.35.0",
    "tenacity>=9.1.2",
    "web3>=7.11.0",
]
